requirement.txt
boto3==1.28.57
python-dotenv==1.0.0
openai==0.29.0
pyodbc==4.0.35
streamlit==1.25.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==1.5.3
plotly==5.16.0
openpyxl==3.1.2
easyocr==1.6.2
requests==2.31.0
beautifulsoup4==4.12.2
pillow==9.5.0
